P4-3191 combining integration tests as the order in which things run needs to be controlled.
